requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
transformers>=4.46.0
sentence-transformers>=3.0.0

# LlamaIndex framework
llama-index>=0.10.53
llama-index-core>=0.10.53
llama-index-embeddings-openai>=0.1.10
llama-index-llms-openai>=0.1.20
llama-index-embeddings-huggingface>=0.2.2

# Embedding and reranking models
FlagEmbedding>=1.2.0

# Machine learning utilities
scikit-learn>=1.5.0
joblib>=1.4.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# NLP utilities
tiktoken>=0.7.0
rouge-score>=0.1.2
nltk>=3.9.0

# Vector search
faiss-cpu>=1.8.0

# API clients
openai>=1.0.0

# Utilities
tqdm>=4.67.0
pydantic>=2.10.0
